Prefer Parliament emails over Constituency ones
Parliamentary contact details are marked as:
  <Address Type_Id="1">

Departmental contacts are:
  <Address Type_Id="2">

Constituency contacts are:
  <Address Type_Id="4">

So ordering by Type_Id befre extracting email addresses is probably the
Right Thing.

(Ideally we'd be able to record which type each email is, but we don't
really have a good model for that atm. Perhaps we should have separate
columns for each, and leave it up to consumers to choose which they
want, and how to combine them, etc?)